Move `CarUseModel` and `TourCombinationModel` into own modules
Improve variable naming
