+++ Scripts/models/logit.py
@@ -4,8 +4,6 @@ import math
 
 from parameters.destination_choice import destination_choice, distance_boundary
 from parameters.mode_choice import mode_choice
-from parameters.car import car_usage
-import parameters.tour_generation as generation_params
 import parameters.zone as zone_params
 from utils.zone_interval import ZoneIntervals
 
@@ -696,256 +694,3 @@ class SecDestModel(LogitModel):
 
 class OriginModel(DestModeModel):
     pass
-
-
-class TourCombinationModel:
-    """Nested logit model for tour combination choice.
-
-    Number of tours per day is the upper level of the model and each
-    number-of-tour nest can have different combinations of tours
-    (e.g., a two-tour combination can be hw-ho, hw-hs or ho-ho, etc.).
-    Base for tour generation.
-
-    Parameters
-    ----------
-    zone_data : ZoneData
-        Data used for all demand calculations
-    """
-
-    def __init__(self, zone_data):
-        self.zone_data = zone_data
-        self.param = generation_params.tour_combinations
-        self.conditions = generation_params.tour_conditions
-        self.increases = generation_params.tour_number_increase
-        self.tour_combinations = []
-        for nr_tours in self.param:
-            self.tour_combinations += self.param[nr_tours].keys()
-    
-    def calc_prob(self, age_group, is_car_user, zones):
-        """Calculate choice probabilities for each tour combination.
-
-        Calculation is done for one specific population group
-        (age + is car user or not) and probabilities are returned for every
-        possible tour combination.
-        
-        Parameters
-        ----------
-        age_group : str
-            Age group (age_7-17/age_18-29/...)
-        is_car_user : bool
-            True if is car user
-        zones : int or slice
-            Zone number (for agent model) or zone data slice
-
-        Returns
-        -------
-        dict
-            key : tuple of str
-                Tour combination (-/hw/hw-ho/...)
-            value : float or numpy 1-d array
-                Choice probability
-        """
-        prob = {}
-        nr_tours_exps = {}
-        nr_tours_expsum = 0
-        for nr_tours in self.param:
-            # Upper level of nested logit model
-            combination_exps = {}
-            combination_expsum = 0
-            for tour_combination in self.param[nr_tours]:
-                # Lower level of nested logit model
-                if tour_combination in self.conditions:
-                    if self.conditions[tour_combination][0]:
-                        # If this tour pattern is exclusively for one age group
-                        if age_group == self.conditions[tour_combination][1]:
-                            is_allowed = True
-                        else:
-                            is_allowed = False
-                    else:
-                        # If one age group is excluded from this tour pattern
-                        if age_group == self.conditions[tour_combination][1]:
-                            is_allowed = False
-                        else:
-                            is_allowed = True
-                else:
-                    is_allowed = True
-                if is_allowed:
-                    param = self.param[nr_tours][tour_combination]
-                    util = 0
-                    util += param["constant"]
-                    for i in param["zone"]:
-                        util += param["zone"][i] * self.zone_data[i][zones]
-                    dummies = param["individual_dummy"]
-                    if age_group in dummies:
-                        util += dummies[age_group]
-                    if is_car_user and "car_users" in dummies:
-                        util += dummies["car_users"]
-                    combination_exps[tour_combination] = numpy.exp(util)
-                else:
-                    combination_exps[tour_combination] = 0
-                combination_expsum += combination_exps[tour_combination]
-            for tour_combination in self.param[nr_tours]:
-                try:
-                    prob[tour_combination] = ( combination_exps[tour_combination]
-                                             / combination_expsum)
-                except ZeroDivisionError:
-                    # Specifically, no 4-tour patterns are allowed for
-                    # 7-17-year-olds, so sum will be zero in this case
-                    prob[tour_combination] = 0
-            util = 0
-            nr_tours_exps[nr_tours] = numpy.exp(util)
-            scale_param = generation_params.tour_number_scale
-            nr_tours_exps[nr_tours] *= numpy.power(combination_expsum, scale_param)
-            nr_tours_expsum += nr_tours_exps[nr_tours]
-        # Probability of no tours at all (empty tuple) is deduced from
-        # other combinations (after calibration)
-        prob[()] = 1
-        for nr_tours in self.param:
-            if nr_tours != 0:
-                nr_tours_prob = nr_tours_exps[nr_tours] / nr_tours_expsum
-                # Tour number probability is calibrated
-                nr_tours_prob *= self.increases[nr_tours]
-                prob[()] -= nr_tours_prob
-                for tour_combination in self.param[nr_tours]:
-                    # Upper and lower level probabilities are combined
-                    prob[tour_combination] *= nr_tours_prob
-        return prob
-
-
-class CarUseModel(LogitModel):
-    """Binary logit model for car use.
-
-    Parameters
-    ----------
-    zone_data : ZoneData
-        Data used for all demand calculations
-    bounds : slice
-        Zone bounds
-    age_groups : tuple
-        tuple
-            int
-                Age intervals
-    resultdata : ResultData
-        Writer object to result directory
-    """
-
-    def __init__(self, zone_data, bounds, age_groups, resultdata):
-        self.resultdata = resultdata
-        self.zone_data = zone_data
-        self.bounds = bounds
-        self.genders = ("female", "male")
-        self.age_groups = age_groups
-        self.param = car_usage
-        for i in self.param["individual_dummy"]:
-            self._check(i)
-    
-    def _check(self, dummy):
-        try:
-            age_interval = dummy.split('_')[1]
-        except AttributeError:
-            # If the dummy is for a compound segment (age + gender)
-            age_interval = dummy[0].split('_')[1]
-            if dummy[1] not in self.genders:
-                raise AttributeError(
-                    "Car use dummy name {} not valid".format(dummy[1]))
-        if tuple(map(int, age_interval.split('-'))) not in self.age_groups:
-            raise AttributeError(
-                "Car use dummy name {} not valid".format(age_interval))
-    
-    def calc_basic_prob(self):
-        """Calculate car user probabilities without individual dummies.
-
-        Returns
-        -------
-        numpy.ndarray
-                Choice probabilities
-        """
-        b = self.param
-        utility = numpy.zeros(self.bounds.stop)
-        self._add_constant(utility, b["constant"])
-        self._add_zone_util(utility, b["generation"], True)
-        self.exps = numpy.exp(utility)
-        self._add_log_zone_util(self.exps, b["log"], True)
-        prob = self.exps / (self.exps+1)
-        return prob
-
-    def calc_prob(self):
-        """Calculate car user probabilities with individual dummies included.
-
-        Returns
-        -------
-        pandas.Series
-                Choice probabilities
-        """
-        prob = self.calc_basic_prob()
-        no_dummy_share = 1
-        dummy_prob = 0
-        b = self.param
-        for i in b["individual_dummy"]:
-            try:
-                dummy_share = self.zone_data.get_data(
-                    "share_"+i, self.bounds, generation=True)
-            except TypeError:
-                # If the dummy is for a compound segment (age + gender)
-                dummy_share = numpy.ones_like(prob)
-                for j in i:
-                    dummy_share *= self.zone_data.get_data(
-                        "share_"+j, self.bounds, generation=True)
-            no_dummy_share -= dummy_share
-            ind_exps = numpy.exp(b["individual_dummy"][i]) * self.exps
-            ind_prob = ind_exps / (ind_exps+1)
-            dummy_prob += dummy_share * ind_prob
-        no_dummy_prob = no_dummy_share * prob
-        prob = no_dummy_prob + dummy_prob
-        prob = pandas.Series(
-            prob, self.zone_data.zone_numbers[self.bounds])
-        self.print_results(prob)
-        return prob
-    
-    def calc_individual_prob(self, age_group, gender, zone=None):
-        """Calculate car user probability with individual dummies included.
-        
-        Uses results from previously run `calc_basic_prob()`.
-        
-        Parameters
-        ----------
-        age_group : str
-            Agent/segment age group
-        gender : str
-            Agent/segment gender (female/male)
-        zone : int (optional)
-            Index of zone where the agent lives, if no zone index is given,
-            calculation is done for all zones
-        
-        Returns
-        -------
-        numpy.ndarray
-                Choice probabilities
-        """
-        self._check((age_group, gender))
-        if zone is None:
-            exp = self.exps
-        else:
-            exp = self.exps[self.zone_data.zone_index(zone)]
-        b = self.param
-        if age_group in b["individual_dummy"]:
-            exp = numpy.exp(b["individual_dummy"][age_group]) * exp
-        if (age_group, gender) in b["individual_dummy"]:
-            exp = numpy.exp(b["individual_dummy"][(age_group, gender)]) * exp
-        prob = exp / (exp+1)
-        return prob
-
-    def print_results(self, prob, population_7_99=None):
-        """ Print results, mainly for calibration purposes"""
-        # Print car user share by zone
-        self.resultdata.print_data(prob, "car_use.txt", "car_use")
-        if population_7_99 is None:
-            # Comparison data has car user shares of population
-            # over 6 years old (from HEHA)
-            population_7_99 = (self.zone_data["population"][self.bounds]
-                               * self.zone_data["share_age_7-99"])
-        # print car use share by municipality and area
-        for area_type in ("municipalities", "areas"):
-            prob_area = ZoneIntervals(area_type).averages(prob, population_7_99)
-            self.resultdata.print_data(
-                prob_area, "car_use_{}.txt".format(area_type), "car_use")
