+++ Scripts/modelsystem.py
@@ -526,7 +526,7 @@ class AgentModelSystem(ModelSystem):
         """
         log.info("Demand calculation started...")
         random.seed(None)
-        self.dm.cm.calc_basic_prob()
+        self.dm.car_use_model.calc_basic_prob()
         for purpose in self.dm.tour_purposes:
             if isinstance(purpose, SecDestPurpose):
                 purpose.init_sums()
@@ -558,7 +558,7 @@ class AgentModelSystem(ModelSystem):
         sec_dest_tours = {mode: [defaultdict(list) for _ in purpose.zone_numbers]
             for mode in purpose.modes}
         car_users = pandas.Series(
-            0, self.zdata_forecast.zone_numbers[self.dm.cm.bounds])
+            0, self.zdata_forecast.zone_numbers[self.dm.car_use_model.bounds])
         for person in self.dm.population:
             person.decide_car_use()
             car_users[person.zone.number] += person.is_car_user
@@ -566,7 +566,7 @@ class AgentModelSystem(ModelSystem):
             for tour in person.tours:
                 tour.choose_mode(person.is_car_user)
                 tour.choose_destination(sec_dest_tours)
-        self.dm.cm.print_results(
+        self.dm.car_use_model.print_results(
             car_users / self.dm.zone_population, self.dm.zone_population)
         log.info("Primary destinations assigned")
         purpose_impedance = self.imptrans.transform(
