+++ Scripts/tests/unit/test_tourcombination.py
@@ -4,7 +4,7 @@ import numpy
 import pandas
 import unittest
 from datahandling.zonedata import ZoneData
-from models.logit import TourCombinationModel
+from models.tour_combinations import TourCombinationModel
 import os
 
 TEST_DATA_PATH = os.path.join(os.path.dirname(os.path.realpath(__file__)), "..", "test_data")
